migrate first distribution test to use simapp
