+++ x/distribution/keeper/common_test.go
@@ -1,0 +1,7 @@
+package keeper_test
+
+import "github.com/cosmos/cosmos-sdk/simapp"
+
+var (
+	PKS = simapp.CreateTestPubKeys(5)
+)
